[cleanup] Fix bad package name
